feat: listen to 2-finger long press and open ECT
To avoid a breaking change, this feature will live side-by-side
with the old code and only run on applications which were built
on MABS 6 and above and thus only target SDK 29 and onwards,
which provides the required version of the Broadcaster plugin.

The breaking change would manifest due to the lack of proper release
segmentation on this asset. Since this plugin always injected at the
latest version of the respective O<version> branch that means it is
possible to have these changes running on an application built with
MABS < 6.

References https://outsystemsrd.atlassian.net/browse/RNMT-3515