fix: do not depend on Broadcaster constants
This would cause a breaking change due to the way this plugin is
(not) versioned. Bear in mind that these constants still mirror
the ones in the Broadcaster plugin and should any of these be
modified there such changes must be reflected here, lest this
code start failing suddenly.

References https://outsystemsrd.atlassian.net/browse/RNMT-3515