fix: view now correctly complies to window insets
When allowing the status bar to overlay the app, the
"navigation bar" of the AppFeedback is not respecting
the window insets and is therefore not being accessible
via touch.

This issue is more serious in notched devices since
the whole bar is being placed below the status bar.
This PR fixes this by forcing the view to comply with
the window insets.

References https://outsystemsrd.atlassian.net/browse/RNMT-3663